+++ Libplanet.Explorer/GraphTypes/BlockType.cs
@@ -33,7 +33,7 @@ namespace Libplanet.Explorer.GraphTypes
             Field(x => x.Timestamp);
             Field<ByteStringType>(
                 "StateRootHash",
-                resolve: ctx => ctx.Source.StateRootHash?.ToByteArray());
+                resolve: ctx => ctx.Source.StateRootHash.ToByteArray());
             Field<NonNullGraphType<ListGraphType<NonNullGraphType<TransactionType<T>>>>>(
                 "transactions"
             );