+++ Libplanet.Tests/Store/StoreFixture.cs
@@ -99,13 +99,13 @@ namespace Libplanet.Tests.Store
                     : (HashDigest<SHA256>?)null;
             GenesisBlock = TestUtils.MineGenesis<DumbAction>(GetHashAlgorithm)
                 .Evaluate(blockAction, stateStore);
-            stateRootHashes[GenesisBlock.Hash] = GenesisBlock.StateRootHash.Value;
+            stateRootHashes[GenesisBlock.Hash] = GenesisBlock.StateRootHash;
             Block1 = TestUtils.MineNextBlock(GenesisBlock, GetHashAlgorithm);
-            stateRootHashes[Block1.Hash] = Block1.StateRootHash.Value;
+            stateRootHashes[Block1.Hash] = Block1.StateRootHash;
             Block2 = TestUtils.MineNextBlock(Block1, GetHashAlgorithm);
-            stateRootHashes[Block2.Hash] = Block2.StateRootHash.Value;
+            stateRootHashes[Block2.Hash] = Block2.StateRootHash;
             Block3 = TestUtils.MineNextBlock(Block2, GetHashAlgorithm);
-            stateRootHashes[Block3.Hash] = Block3.StateRootHash.Value;
+            stateRootHashes[Block3.Hash] = Block3.StateRootHash;
 
             Transaction1 = MakeTransaction(new List<DumbAction>(), ImmutableHashSet<Address>.Empty);
             Transaction2 = MakeTransaction(new List<DumbAction>(), ImmutableHashSet<Address>.Empty);